chore: remove moment
