+++ package.json
@@ -84,8 +84,6 @@
     "humanize-duration": "3.28.0",
     "lodash-es": "4.17.21",
     "luxon": "3.2.1",
-    "moment": "2.29.4",
-    "moment-timezone": "0.5.39",
     "object-hash": "2.2.0",
     "rxjs": "^7.5.7",
     "scroll-into-view-if-needed": "2.2.28",
@@ -118,7 +116,6 @@
     "@types/jasminewd2": "2.0.6",
     "@types/lodash-es": "4.17.4",
     "@types/luxon": "3.2.0",
-    "@types/moment-timezone": "0.5.30",
     "@types/object-hash": "2.1.0",
     "@types/uuid": "^9.0.0",
     "@typescript-eslint/eslint-plugin": "^5.44.0",
